Refactor into ItemsScroll.
